Extract :scissors: to DefaultBody
