+++ fuel/src/main/kotlin/com/github/kittinunf/fuel/core/requests/DefaultRequest.kt
@@ -3,7 +3,6 @@ package com.github.kittinunf.fuel.core.requests
 import com.github.kittinunf.fuel.core.Body
 import com.github.kittinunf.fuel.core.BodyLength
 import com.github.kittinunf.fuel.core.BodySource
-import com.github.kittinunf.fuel.core.DefaultBody
 import com.github.kittinunf.fuel.core.Handler
 import com.github.kittinunf.fuel.core.HeaderValues
 import com.github.kittinunf.fuel.core.Headers
@@ -29,12 +28,12 @@ import java.net.URL
 import java.nio.charset.Charset
 
 data class DefaultRequest(
-    override val method: Method,
-    override val url: URL,
-    override val headers: Headers = Headers(),
-    override val parameters: Parameters = listOf(),
-    internal var _body: Body = DefaultBody(),
-    override val enabledFeatures: RequestFeatures = mutableMapOf()
+        override val method: Method,
+        override val url: URL,
+        override val headers: Headers = Headers(),
+        override val parameters: Parameters = listOf(),
+        internal var _body: Body = DefaultBody(),
+        override val enabledFeatures: RequestFeatures = mutableMapOf()
 ) : Request {
 
     override lateinit var executionOptions: RequestExecutionOptions
